move generated clients to pkg/clients/
