+++ pkg/clients/clientset/clientset.go
@@ -19,7 +19,7 @@ limitations under the License.
 package clientset
 
 import (
-	migrationv1alpha1 "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clientset/typed/migration/v1alpha1"
+	migrationv1alpha1 "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clients/clientset/typed/migration/v1alpha1"
 	discovery "k8s.io/client-go/discovery"
 	rest "k8s.io/client-go/rest"
 	flowcontrol "k8s.io/client-go/util/flowcontrol"
