+++ pkg/clients/clientset/fake/register.go
