+++ pkg/clients/clientset/scheme/register.go
