+++ pkg/clients/clientset/typed/migration/v1alpha1/fake/fake_migration_client.go
@@ -19,7 +19,7 @@ limitations under the License.
 package fake
 
 import (
-	v1alpha1 "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clientset/typed/migration/v1alpha1"
+	v1alpha1 "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clients/clientset/typed/migration/v1alpha1"
 	rest "k8s.io/client-go/rest"
 	testing "k8s.io/client-go/testing"
 )
