+++ pkg/clients/clientset/typed/migration/v1alpha1/storageversionmigration.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	v1alpha1 "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/apis/migration/v1alpha1"
-	scheme "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clientset/scheme"
+	scheme "github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clients/clientset/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
 	watch "k8s.io/apimachinery/pkg/watch"
