+++ pkg/trigger/discovery_handler_test.go
@@ -23,7 +23,7 @@ import (
 	"testing"
 
 	"github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/apis/migration/v1alpha1"
-	"github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clientset/fake"
+	"github.com/kubernetes-sigs/kube-storage-version-migrator/pkg/clients/clientset/fake"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
